something wrong in CHILD subscription model: don't send events directly to parent, always publish

+++ kernel/c/root-esp32/components/esp_yuneta/src/c_esp_transport.c
@@ -226,9 +226,9 @@ PRIVATE void mt_create(hgobj gobj)
     gobj_subscribe_event(gobj_yuno(), EV_TIMEOUT_PERIODIC, NULL, gobj);
 
     /*
-     *  CHILD subscription model TODO check
+     *  CHILD subscription model
      */
-    hgobj subscriber = (hgobj)(size_t)gobj_read_integer_attr(gobj, "subscriber");
+    hgobj subscriber = (hgobj)gobj_read_integer_attr(gobj, "subscriber");
     if(!subscriber) {
         subscriber = gobj_parent(gobj);
     }
