+++ kernel/c/root-esp32/components/esp_yuneta/src/c_esp_uart.c
@@ -127,9 +127,9 @@ PRIVATE void mt_create(hgobj gobj)
 #endif
 
     /*
-     *  CHILD subscription model TODO check
+     *  CHILD subscription model
      */
-    hgobj subscriber = (hgobj)(size_t)gobj_read_integer_attr(gobj, "subscriber");
+    hgobj subscriber = (hgobj)gobj_read_integer_attr(gobj, "subscriber");
     if(!subscriber) {
         subscriber = gobj_parent(gobj);
     }
