+++ kernel/c/root-linux/src/c_uart.c
@@ -126,7 +126,7 @@ PRIVATE void mt_create(hgobj gobj)
     /*
      *  CHILD subscription model
      */
-    hgobj subscriber = (hgobj)(size_t)gobj_read_integer_attr(gobj, "subscriber");
+    hgobj subscriber = (hgobj)gobj_read_integer_attr(gobj, "subscriber");
     if(!subscriber) {
         subscriber = gobj_parent(gobj);
     }
